requirements.txt
EWMHlib==0.2
MouseInfo==0.1.3
Pillow==9.5.0
pkg_resources==0.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMonCtl==0.92
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
PyWinBox==0.7
PyWinCtl==0.4
six==1.17.0
typing_extensions==4.7.1
